third_party/blink/web_tests/fast/block/float/float-in-float-hit-testing-expected.txt

CONSOLE MESSAGE: A soft navigation has been detected: float-in-float-hit-testing.html#
